backend/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distlib==0.3.8
dotenv==0.9.9
filelock==3.15.4
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
limits==4.7.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
pipenv==2024.0.1
platformdirs==4.2.2
psutil==7.0.0
psycopg2-binary==2.9.10
pyodbc==5.2.0
python-dotenv==1.0.1
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.1
# redis==5.2.1  # DISABLED - Not using Redis, using memory storage
# Dependencias principales - Base de datos IAMC (SQL Server)
pyodbc>=5.0.0
python-dotenv>=1.0.0
# psycopg2>=2.9.0  # LEGACY - Ya no se usa PostgreSQL, migrado a SQL Server IAMC
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
setuptools==72.2.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.39
sqlparse==0.5.1
sympy==1.13.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
pywin32==308
weasyprint==62.3